Use GNU C __attribute__
The new header defines macros that correspond to GCC function
attributes: `PURE`, `NOTNULL(...)`, and `PRINTF(fmt, arg1)`.
An empty definition of the builtin function `__attribute__(x)` is made
if the compilation is in ISO C mode so that the compiler does not
complain.

The attributes are added to the declaration of the suitable functions.