Fix blank messages behaviour
A blank message could have been output whatever its level, even if it
should be filtered out.
This behaviour is changed and the message will now be filtered out if
its priority level is not high enough.

Braces are added around single-statement `if` branches.
The `fmt` argument of `vlogmsg` is not `const` anymore, because its
value can be incremented to exclude a leading newline, if present.