Output attributes
A new enum defines values that specify attributes to modify the log
display with.
The zero value is the minimal output -- the same as before. A value
activates the output of the time, in format `hh:mm:ss`, another prints
the file name and line number, yet another displays the name of the
function the log was called from.
The last attribute enables colored display.

Each priority level has an associated color that makes its output -- if
enabled -- more user-friendly.

Two functions are added to set and retrieve the output attributes.

The logging functions `logmsg` and `vlogmsg` are added parameters: file
name, line number, function name; those values should be the evaluation
of the standard macros `__FILE__`, `__LINE__` and `__func__`,
respectively. Therefore it is more advised to call the usual logging
macros onamed after their level `debug`, `warning`, etc. that mention
the macros themselves.