Add INLINE attribute and use in static funcs
The `INLINE` attribute forces a function to be inlined, even without
optimization enabled; using it allows to refactor code but without
performance loss.
The internal functions `_msgblank` and `_printtime` are declared with
the attribute.
The function `_msgblank` had to be rewritten in an iterative form,
recursive funcs not being inlinable.