Add DTD
The DTD defines the log structure for the XML output format.

There is a root element, `log`, that can contain a (non-zero) number of
`msg` elements. This element contains only text data: the message
content. It also has a number of attributes: `time, the time of logging
(optional), `file`, the name of the file (optional), `line`, the line
number inside the file (optional) and `level`, the priority level of
the message. This attribute is required.