Add JSON output format
In this format, the log are output in pretty JSON (with whitespace).
This format is incompatible with CLOG_INIT_APPEND since there is
terminating output (closing braces and brackets).

The output in this format resembles:

    {
        "log": [
            {
                "time": "hh:mm:ss",
                "file": "myfile.c",
                "line": 42,
                "func": "main",
                "level": "warning",
                "msg": "There is a bug!"
            }
         ]
    }

Trailing commas are not allowed in JSON arrays and objects, so the
first-element case in gracefully handled with an internal boolean; when
`true`, the comma separating this message from the previous is not
output.