Remove append mode for log output
The *append output to end of file* feature is removed; it was not
supported by all output formats (it would have broken XML or JSON
syntax) and would only result in an error of the initializer function,
or could be found problematic (what if one appended log with fewer
attributes, ie less columns, to a CSV file?) so this feature did not
even make sense.

The `InitMode` enumeration is completely removed since it would consist
of only one constant. All references of it, `CLOG_INIT_TRUNCATE` and
`CLOG_INIT_APPEND` are removed; the array that matched the constants
with the parameter to `fopen` is also removed and the string hard-coded
in place of the array access, in the initializer function
`clog_init_file`.