Add TRACE level, with lower priority than DEBUG
The purpose of this new level is to mark the state of the control flow,
for example when entering or leaving the function.

It is a specialization of the level *DEBUG* which should now rather be
used for other messages, such as outputting the values of variables, or
any development (temporary) message.