Export the installation directory in Makefile
The variable `INST_DIR` is exported so that it redefines that used in
the submodule. It allows to centralize the definition of the install
directory instead of having all dependencies scattered throughout the
filesystem--this would be unlikely to happen anyway as long as I'm the
only dev on these projects but it never hurts being rigorous.